requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dateutil==2.9.0.post0
pandas==2.2.2
MetaTrader5==5.0.5260
requests
python-dotenv
